Get dummie lists
* requestDesc.txt
Text file that describe the main functionalities that should be implemented.

* golist.go
- Define the list type, which is composed by an ID and items.
Items are composed by a name and boolean that represent if it has been
checked or not.
- Create a function that returns a dummy list of lists.
- Now handle access to /lists, where the list of lists is printed (both at
the page and at the terminal)

Interesting notes:
	We can annotate variables to inform which name they should take when
converted to json.
	Go treats variables in a structure starting by an upper case letter
as public and a lower case as private. Private variables are ignored when
the structure is being encoded to json. (I lost quite the amount of time trying
to figure out why some variables were being ignored).